[Update] HP→ Heart, MP→Stamina 수정하기 float HP → int Heart 로 관련된 부분 수정하기

+++ Assets/ScriptableObject/Script/Item/ConsumableItemData.cs
@@ -11,6 +11,6 @@ public class ConsumableItemData : ItemData
 {
     [Header("Consum Info")]
     public ConsumableItemType type;
-    public float increaseValue;
+    public int increaseValue;
 }
